add token hook
