+++ template/src/SDKAppWrapper.tsx
@@ -1,6 +1,5 @@
 import React, {useEffect, useState} from 'react';
 import {CustomizationApiInterface, customize} from 'customization-api';
-import AsyncStorage from '@react-native-community/async-storage';
 import {
   customizationConfig,
   CustomizationProvider,
@@ -8,9 +7,7 @@ import {
 import SDKEvents from './utils/SdkEvents';
 import App from './App';
 import {StyleSheet, View, Text, ActivityIndicator} from 'react-native';
-import jwt_decode from 'jwt-decode';
-
-const REFRESH_TOKEN_DURATION_IN_SEC = 60;
+import useTokenServerAuth from './auth/useTokenServerAuth';
 
 export interface userEventsMapInterface {
   leave: () => void;
@@ -69,87 +66,25 @@ export const AppBuilderSdkApi: AppBuilderSdkApiInterface = {
 
 const SDKAppWrapper = () => {
   const [fpe, setFpe] = useState(customizationConfig);
+  const {setServerToken, validateToken} = useTokenServerAuth();
   const [initialized, setInitialized] = React.useState(false);
-  const [tokenExpiresAt, setTokenExpiresAt] = React.useState(0);
-  const [sdkToken, setSdkToken] = React.useState(null);
-
-  const enableSdkAuthentication = (options: OptionsInterface) => {
-    if (options && options.token && options.token.trim() !== '') {
-      // console.log('1. supriya enabling authentication', options.token);
-      setSdkToken(options.token);
-    } else {
-      throw 'Token is missing in the options';
-    }
-  };
-
-  const getNewToken = async () => {
-    await fetch(`${$config.BACKEND_ENDPOINT}/token/refresh`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: sdkToken ? `Bearer ${sdkToken}` : '',
-      },
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        // console.log('7. supriya get new token', data);
-        setSdkToken(data.token || null);
-      });
-  };
+  const [options, setSdkOptions] = React.useState<OptionsInterface>();
 
   useEffect(() => {
-    const syncSdk = async () => {
-      if (!sdkToken) return;
-      // console.log('2. supriya sdk token set', sdkToken);
-      const decoded = jwt_decode(sdkToken);
-      const expiresAt = decoded?.exp * 1000;
-      if (Date.now() >= expiresAt) {
-        throw 'Token expired. Pass a new token';
-      } else {
-        // console.log('3. supriya emitting token');
-        SDKEvents.emit('sdk-token', sdkToken);
-        setTokenExpiresAt(expiresAt);
-      }
-    };
-    initialized && syncSdk();
-  }, [initialized, sdkToken]);
-
-  useEffect(() => {
-    if (!tokenExpiresAt) return;
-
-    const timer = setInterval(() => {
-      const diffInSeconds = Math.floor(
-        Math.abs(tokenExpiresAt - Date.now()) / 1000,
-      );
-      // console.log('4. supriya diff: ', diffInSeconds);
-      // const minutesLeft = Math.floor(diffInSeconds / 60) % 60;
-      // const secondsLeft = (diffInSeconds - minutes * 60) % 60;
-
-      if (diffInSeconds < REFRESH_TOKEN_DURATION_IN_SEC) {
-        // console.log('5 . supriya 1 min range reached');
-        try {
-          getNewToken();
-          clearInterval(timer);
-        } catch (error) {
-          clearInterval(timer);
-        }
-      }
-    }, 1000);
-
-    return () => {
-      clearInterval(timer);
-    };
-  }, [tokenExpiresAt]);
+    if (!initialized) return;
+    $config.ENABLE_TOKEN_SERVER && setServerToken(options.token);
+  }, [initialized]);
 
   useEffect(() => {
     SDKEvents.on('appInit', async (options, resolve, reject) => {
       try {
-        $config.ENABLE_SDK_AUTHENTICATION && enableSdkAuthentication(options);
+        $config.ENABLE_TOKEN_SERVER && validateToken(options);
+        setSdkOptions(options);
         setInitialized(true);
         resolve(true);
       } catch (error) {
         setInitialized(false);
-        reject(`Error initializing app buider ${error}`);
+        reject(`Error initializing app builder ${error}`);
       }
     });
     SDKEvents.on('addFpe', (sdkFpeConfig) => {
@@ -166,13 +101,14 @@ const SDKAppWrapper = () => {
           <App />
         </CustomizationProvider>
       ) : (
-        <View style={styles.overlay}>
-          <Text style={styles.loadingText}>Initializing app...</Text>
-          <ActivityIndicator
-            size="large"
-            color={$config.SECONDARY_FONT_COLOR}
-          />
-        </View>
+        <></>
+        // <View style={styles.overlay}>
+        // <Text style={styles.loadingText}>Initializing app...</Text>
+        // <ActivityIndicator
+        //   size="large"
+        //   color={$config.SECONDARY_FONT_COLOR}
+        // />
+        // </View>
       )}
     </>
   );
